Added speech recognition thread
This commit adds speech recognition functionality to the
assistant mode. A grammar file has been added to increase
accuracy. There was a issue with the boolean check that kills
the assistant thread not being updated when the user selects the
menu button so the boolean was changed to a volatile bool so
that the thread should be killed when the user exits the assistant mode.