combine all steps into one class for hibernate
In order to avoid issues with loading the macros using hibernate,
this commit combines all step child classes into a single parent
Step class. Though I realize this is less cohesive in terms of
design, I have not been able to find a solution in hibernate
to work with my particular case, so this will have to work
for now. Macros can now be successfully saved and loaded with the list
of steps attached.