fixed issue with speechHandler constructor
This commit fixes a bug I introduced when I removed
the 'this' keyword from the constructor which prevented
the speech interpreter from being properly set. The
variable name for the interpreter that is passed in to
the constructor has been changed so it is different than
the member var name.