added checks for creating var steps
This commit adds logic checking to verify that
a user cannot finish a var step before starting one.
Also added check to verify that a user cannot finish
making a macro while they are in the middle of making
a variable step.

Also added more phrases for assistant mode to help give users
options to turn off pamm in case the accuracy fails on other commands.