Using streams instead of files to make compatible with JAR
Before this commit, whenever I attempted to package the project
in a jar it could not run because the program attempts to update
the grammar file at runtime. This commit is the first step to making
the grammar file external to the jar by using streams to read the
starting grammar file from resources and writing it to another location.
The media player was also updated to use a URI to the sound file rather
than creating a file object.