Added speech handling fail safe
Because of issues experienced with speech recognition accuracy,
this commit adds a fail safe mechanism that pops up the list of user
created commands if the program does not recognize 3 commands in a row.
The user can then manually run the command from the list menu.

Also added a tooltip on screen which displays a list of system commands
when clicked. This dialog uses the same view file as the custom command
list.