Added custom speech recognizer
The LiveSpeechRecognizer class from Sphinx could not create more
than one instance of microphone without crashing on Windows.
This prevented me from being able to update the configuration of
the grammar at runtime so I created a customSpeechRecognizer that
is identical except the microphone is static and only initialized once.
This seems to allow updating the configuration without crashing.