running updates in separate thread
Whenever the grammar is updated, the recognizer is now updated
in a separate thread so the UI is not hung up waiting. Added a
completable future that allows the program to make sure the speech
recognition is ready before listening.

Also fixed issue with db deletion method being static, this was causing
the db to lock up so it was changed to non-static.