Making sure microphone is cleared
There was an issue where the assistant mode can be closed with the
microphone running. If this happens the user must clear the microphone
by making an utterance before using the microphone on another thread.
This commit adds a completable future object that checks for this
case and if the microphone is not cleared, a dialog window is displayed
to the user prompting them to say something before the macro create mode
is run.