Added error message if user presses multimedia key
To prevent possible crashing of the java robot class not being
able to define media keys, this adds a check to typing actions to
verify the key code is defined. If it is not an error message pops
up to alert the user that their key press is an invalid action. I'm
sure there are edge cases not covered by this check but for now
this should help prevent recording of invalid key presses.