checking for undefined keys
This commit adds a check to all key typing events to verify that the
key typed by the user is able to be typed by the robot class. Found
that the value sent by the right shift button is not recognized by the
KeyEvent class for some reason, so we filter that from the error messages.

Also added error handling in the typeKeyCommand method to make sure keys
are released and the user is notified if an error occurs