covered recording exit point
if the user gets stuck in the recording view for whatever reason and
attempts the forcefully close the kill switch window, the program now
returns them to the main menu and stops all recording tasks. Also added
System.exit call when user attempts to close the program to make sure
the JVM stops running.